+++ lib/src/list_extensions.dart
@@ -1,5 +1,23 @@
+import 'dart:math';
+
 /// Extension methods for any [List].
 extension ListExtensions<E> on List<E> {
+  // Common
+
+  // Common - Get
+
+  /// Returns a random element from the list.
+  ///
+  /// Throws a [StateError] if `this` is empty.
+  E get random {
+    if (isEmpty) {
+      throw StateError('No element');
+    }
+
+    final rnd = Random();
+    return this[rnd.nextInt(length)];
+  }
+
   // Transformation
 
   // Transformation - List