+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Next release
 
+
+* Getter `random` - returns a random element from the list.
+
 ## [0.1.11+1] - 2020-07-16
 
 * Readme fix.